changing to 32bit indexes per vector

+++ frf.h
@@ -13,13 +13,13 @@
 typedef struct frf {
   int        num_rows;
 
-  uint16_t * _vector_base;
+  uint32_t * _vector_base;
   uint32_t * _vector_header_base;
   char     * _string_table_base;
   uint32_t * _mmap_base;
 
-  uint16_t * _row_ptr;
-  uint16_t * _end;
+  uint32_t * _row_ptr;
+  uint32_t * _end;
 } frf_t;
 
 int frf_init(frf_t * frf, char * file_name);
